added some aditional validation funtionality for the statistic files

+++ modules/utilities.rb
@@ -157,8 +157,15 @@ module Sinatra
     def save_statfile(params)
       errors = {}
       filename = h params[:dsname].to_s.tr(' ', '_')
+      unless Dir.exists?('statisticfiles/')
+        begin
+          Dir.mkdir('statisticfiles/')
+        rescue SystemCallError => e
+          errors[:mkdir] = "can't create statistics directory" 
+        end
+      end
       filepath = 'statisticfiles/'+filename
-      if (!File.exists?(filepath)||  params[:overwrite])
+      if ( errors.empty? && (!File.exists?(filepath)||  params[:overwrite]))
         File.open(filepath, 'w') do |statfile|
           statfile.write(params[:statfile][:tempfile].read)
         end
